Refactors some of the code
