+++ app/src/main/java/com/smd/surmaiya/ManagerClasses/MusicServiceManager.kt
@@ -7,7 +7,9 @@ import android.content.ServiceConnection
 import android.graphics.Bitmap
 import android.net.Uri
 import android.os.Build
+import android.os.Handler
 import android.os.IBinder
+import android.os.Looper
 import android.util.Log
 import androidx.annotation.RequiresApi
 import com.smd.surmaiya.Services.MusicService
@@ -64,22 +66,38 @@ object MusicServiceManager {
     }
 
     @RequiresApi(Build.VERSION_CODES.P)
-    fun pauseSong() {
-        songManager.currentProgress = (musicService?.getProgress() ?: 0).toFloat()
-        musicService?.pauseMusic()
+    fun playCurrentSongWithDelay(delayMillis: Long) {
+        val progress = musicService?.getProgress()?.toFloat() ?: 0f
+        SongManager.getInstance().currentSong?.let { song ->
+            playMusicWithDelay(song, progress, delayMillis)
+        }
+    }
+
+    @RequiresApi(Build.VERSION_CODES.P)
+    fun playMusicWithDelay(song: Song, progress: Float, delayMillis: Long) {
+        getService()?.playSong(song, progress)
+
+        Handler(Looper.getMainLooper()).postDelayed({
+            val duration = getService()?.exoPlayer?.duration ?: 0
+            Log.d("Progress percentage= ", "Progress percentage = " + progress/100)
+            Log.d("DUration = ", "DUration = " + duration)
+            Log.d("Progress = ", "Progress = " + progress)
+            val newProgress = (progress/ 100.0) * duration
+            getService()?.exoPlayer?.seekTo(newProgress.toLong())
+        }, delayMillis) // Delay of 1 second
+
+        val intent = Intent("com.smd.surmaiya.ACTION_PLAY")
+        musicService?.sendBroadcast(intent)
+    }
+
+    fun pauseMusicAndBroadcast() {
+        musicService?.progress = musicService?.getProgress()?.toFloat() ?: 0f
+        musicService?.exoPlayer?.stop()
 
         val intent = Intent("com.smd.surmaiya.ACTION_PAUSE")
         musicService?.sendBroadcast(intent)
     }
 
-    fun resumeSong() {
-
-        songManager.currentProgress = (musicService?.getProgress() ?: 0).toFloat()
-        musicService?.resumeSong()
-
-        val intent = Intent("com.smd.surmaiya.ACTION_PLAY")
-        musicService?.sendBroadcast(intent)
-    }
 
     fun getService(): MusicService? {
         return musicService
